I took the bt_model and modified it slightly to solve the surface quasi-geostrophic model (sqg).
Note that the inversion that I use between the buoyancy and the streamfunction assumes an infinite bottom.

I also introduced a buoyancy frequency, Nb, but it is set to 1.0 so it does nothing as of now.

I have some latex notes that I can borrow from to have a proper description of this in the documentation.  After I figure out how sphinx works of course.