Remove automatic OpenMP detection.
This is difficult to do in general and complicates the build since in
many cases (i.e. uploading to PyPI) we would need to suppress OpenMP
support in any case to avoid linking to the library. Likely the
simpler approach is to enable it when needed by adding "-fopenmp" to
CFLAGS when building, for example on conda-forge.